refactor(bus): convert area autocomplete to typescript
- replaced `AreaAutocomplete.js` with `AreaAutocomplete.tsx` because new BUS components should have followed the team's `TypeScript` convention and the existing `Bus` domain types were already available
- added explicit prop and helper argument types because the autocomplete flow relied on structured area selection data and the type contracts made the component API clearer
- removed `propTypes` because the new `TypeScript` component now expressed the runtime shape expectations directly in the source
- documented the `setTimeout(..., 0)` workaround in the selection sync effect because it guarded a concrete `Autocomplete` and `SearchBar` UI bug where stale native text could overwrite the selected label after an external area change
